Update to new Localization and Cleanup
- Change: Moved DailyClaimResponse into it's own file.
- Change: DailyService and DailyClaimResponse are in their own namespace now.
- Add: Daily localization strings into eng.json.
