+++ test/Miki.Tests/Services/DailyServiceTests.cs
@@ -4,6 +4,7 @@
     using System.Threading.Tasks;
     using Miki.Bot.Models;
     using Miki.Services;
+    using Miki.Services.Daily;
     using Miki.Services.Transactions;
     using Xunit;
